refactor(runtime): extract turn/stream_guard.rs into agent/turn — #7415
