+++ crates/zeroclaw-runtime/src/agent/turn/mod.rs
@@ -10,14 +10,11 @@ pub(crate) mod events;
 pub(crate) mod outcome;
 pub(crate) mod protocol_detect;
 pub(crate) mod redact;
+pub(crate) mod stream_guard;
 
 pub(crate) use delivery_defaults::maybe_inject_channel_delivery_defaults;
 pub(crate) use protocol_detect::{
-    complete_json_fence_protocol_state, complete_non_protocol_json,
     detect_internal_protocol_without_tools, detect_tool_call_parse_issue_for_known_tools,
-    find_embedded_protocol_candidate_start, find_incomplete_protocol_candidate_start,
-    longest_suffix_matching_prefix, starts_suspicious_protocol_prefix,
-    starts_suspicious_tag_or_fence_prefix,
 };
 pub use events::{DraftEvent, PROGRESS_MIN_INTERVAL_MS, StreamDelta};
 pub(crate) use events::STREAM_CHUNK_MIN_CHARS;
@@ -26,3 +23,4 @@ pub use outcome::{
     is_tool_loop_cancelled,
 };
 pub use redact::scrub_credentials;
+pub(crate) use stream_guard::{StreamTextGuard, StreamThinkTagStripper};
